Adding more Apple Music Types && Cleaning up GetSpotifyMedia function
